#138 Rename makeSimpleActionCreator to makePayloadActionCreator
