+++ CHANGELOG.md
@@ -14,8 +14,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
+- (_actions_): `makeSimpleActionCreator` has been renamed to `makePayloadActionCreator`.
 - (_actions_): `makeActionCreator` has been renamed to `configureActionCreator`.
-- (_actions_): Changed signatures of `makeSimpleActionCreator` and `makeConstantActionCreator`.
+- (_actions_): Changed signatures of `makePayloadActionCreator` and `makeConstantActionCreator`.
 - (_injectors-react_): `Provider` has been renamed to `NamespaceProvider`.
 - (_actions_): `makeReducer` has been moved to the `@redux-tools/reducers` package.
 - (_reducers_): `getStateByAction` and `getStateByNamespace` have been moved to the `@redux-tools/namespaces` package.
