+++ packages/actions/src/configureActionCreator.js
@@ -14,7 +14,7 @@ const isUndefined = value => value === undefined;
  * @example
  *
  *    const reset = makeConstantActionCreator("RESET")
- *    const add = makeSimpleActionCreator("ADD");
+ *    const add = makePayloadActionCreator("ADD");
  *    const fetchItems = configureActionCreator("FETCH_ITEMS", R.prop("items"), R.always({ page: 0 }))
  */
