+++ packages/actions/src/index.js
@@ -3,7 +3,7 @@ export { prefixedValueMirror as makeActionTypes } from '@redux-tools/utils';
 export { default as configureActionCreator } from './configureActionCreator';
 export { default as makePayloadMetaActionCreator } from './makePayloadMetaActionCreator';
 export { default as makeConstantActionCreator } from './makeConstantActionCreator';
-export { default as makeSimpleActionCreator } from './makeSimpleActionCreator';
+export { default as makePayloadActionCreator } from './makePayloadActionCreator';
 
 export { default as isErrorAction } from './isErrorAction';
 export { default as isNotErrorAction } from './isNotErrorAction';
