+++ packages/actions/src/makePayloadActionCreator.test.js
@@ -1,21 +1,21 @@
-import makeSimpleActionCreator from './makeSimpleActionCreator';
+import makePayloadActionCreator from './makePayloadActionCreator';
 
-describe('makeSimpleActionCreator', () => {
+describe('makePayloadActionCreator', () => {
 	it('throws when a type is empty string', () => {
-		expect(() => makeSimpleActionCreator('')).toThrow();
+		expect(() => makePayloadActionCreator('')).toThrow();
 	});
 
 	it('throws when a type is empty', () => {
-		expect(() => makeSimpleActionCreator()).toThrow();
+		expect(() => makePayloadActionCreator()).toThrow();
 	});
 
 	it('uses the argument as the action payload', () => {
-		const actionCreator = makeSimpleActionCreator('TYPE');
+		const actionCreator = makePayloadActionCreator('TYPE');
 		expect(actionCreator('yo')).toEqual({ type: 'TYPE', payload: 'yo' });
 	});
 
 	it('throws when an undefined payload is received', () => {
-		const actionCreator = makeSimpleActionCreator('TYPE');
+		const actionCreator = makePayloadActionCreator('TYPE');
 		expect(() => actionCreator()).toThrow();
 	});
 });
